+++ NEWS.md
@@ -1,5 +1,4 @@
-# SampleSizeU (development version)
-
 # SampleSizeU 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added the function rate_power